move page and nav types
