+++ modules/common/nav/gameIndexLink.ts
@@ -1,4 +1,4 @@
-import { AlgolNavLink } from "../../ui/src/helpers";
+import { AlgolNavLink } from "../../types";
 
 export const gameIndexLink: AlgolNavLink = {
   title: "Games",
