+++ modules/common/nav/homeNav.ts
@@ -1,4 +1,4 @@
-import { AlgolNav } from "../../ui/src/helpers";
+import { AlgolNav } from "../../types";
 import { homeLink } from "./homeLink";
 import { gameIndexLink } from "./gameIndexLink";
 import { tagIndexLink } from "./tagIndexLink";
