+++ modules/common/nav/makeGameAboutNav.ts
@@ -1,5 +1,4 @@
-import { AlgolGameBlobAnon, AlgolMeta } from "../../types";
-import { AlgolNav } from "../../ui/src/helpers";
+import { AlgolGameBlobAnon, AlgolMeta, AlgolNav } from "../../types";
 import { makeGameAboutLink } from "./makeGameAboutLink";
 import { homeLink } from "./homeLink";
 import { gameIndexLink } from "./gameIndexLink";
