+++ modules/common/nav/makeTagArticleNav.ts
@@ -1,5 +1,4 @@
-import { AlgolArticle } from "../../types";
-import { AlgolNav } from "../../ui/src/helpers";
+import { AlgolArticle, AlgolNav } from "../../types";
 import { homeLink } from "./homeLink";
 import { tagIndexLink } from "./tagIndexLink";
 import { makeTagArticleLink } from "./makeTagArticleLink";
