+++ modules/types/index.ts
@@ -11,3 +11,4 @@ export * from "./blob";
 export * from "./misc";
 export * from "./content";
 export * from "./payloads";
+export * from './page'
