+++ modules/types/page/AlgolPage.ts
@@ -1,5 +1,8 @@
-import { AppActions, BattleNavActions, BattleMode, AlgolNav } from ".";
 import { FunctionComponent } from "react";
+
+import { AppActions } from './appActions'
+import { BattleMode, BattleNavActions } from './battleActions'
+import { AlgolNav } from './nav'
 
 export type AlgolPageProps = {
   actions: AppActions & BattleNavActions;
