+++ modules/types/page/index.ts
@@ -1,0 +1,4 @@
+export * from './AlgolPage'
+export * from './appActions'
+export * from './battleActions'
+export * from './nav'
