+++ modules/ui/src/helpers/index.ts
@@ -1,7 +1,4 @@
 export * from "./positionStyles";
 export * from "./useDemo";
-export * from "./appActions";
 export * from "./useModal";
 export * from "./copy";
-export * from "./battleActions";
-export * from "./nav";
