+++ modules/ui/src/helpers/usePrefetchList.ts
@@ -1,5 +1,5 @@
 import { useRef, useEffect } from "react";
-import { AppActions } from "./appActions";
+import { AppActions } from "../../../types";
 
 export const usePrefetchList = (actions: AppActions, urls: string[]) => {
   const interval = useRef<ReturnType<typeof setInterval> | null>(null);
